add the crew module to the falcon9 system
This adds the crew package to the falcon9 project. This package includes an interface used to represent a crew member. In addition, it includes an implementation of that interface.

Each crew member consists of their name (for display purposes) and a unique key (for identifying purposes).

In addition to the crew module, a `LICENSE` file and a `.travis.yml` file were added.