add missions to falcon9
This change adds the initial implementation of a mission. This does not complete the functionality of a mission, but seems to be a good starting point.

This change adds the initial interfaces used by a mission. That includes the interface for managing crew members, but also the interface for inspecting the parameters of a mission. In addition to the functionality being added, tests were written to confirm the new functionality.