add support for aborting a mission
This change converts the `f9mission.Mission` struct to use an `fsm.Machine` struct to maintain mission state. By using this state machine, we can be a bit more fine-grained with the states of the mission.

Once we had the state machine, it made it easier to implement the idea of aborting a blastoff. So we add the `VoteAbort` vote type, and implement the logic around aborting a mission.

The logic that was added is being tested in the unit tests.