convert mission ID from a string to a uint32
The mission ID is going to be sent over the wire. Initially, I was going to use a UUID or something, but that is going to require a lot more data be sent across the wire. Instead, let's use four bytes (uint32) to generate a pseudo-random identification number. `uint32` should contain enough values for the server to operate without collisions.